Fix Travis'
